Some more cleanup.
