+++ packages/gramophone/src/setup/server.ts
@@ -2,14 +2,9 @@ import fastify from 'fastify';
 import cors from 'fastify-cors';
 import { setup as sockets } from 'sockets';
 import * as endpoints from 'endpoints';
-import { auth } from 'middleware/auth';
 
 export default () => {
   const server = fastify({ logger: true });
-  server.decorate('auth', () => {
-    auth(server);
-    return server;
-  });
   server.register(cors);
 
   // Health check endpoint.
